Use unix time stamps instead of Instant
kotlinx.datetime under the hood calls some Java APIs not available on Android API 25 and lower. This forces clients using this library to use desugaring.
We don't wan't that. And since time calculations are not very crucial in this project, we just use seconds since epoch instead, which can just be a Long.

+++ auth/src/main/kotlin/com/tidal/sdk/auth/model/UpgradeResponse.kt
@@ -12,7 +12,7 @@ internal data class UpgradeResponse(
     @JsonNames("token_type")
     val tokenType: String?,
     @JsonNames("expires_in")
-    val expiresIn: Int? = 0,
+    val expiresIn: Long? = 0,
     @JsonNames("user_id")
     val userId: Int? = null,
 )
